refactor(auth): move gql statements to hooks and abstracts mutation logic
