--- web/src/constants/errorMessages.ts
@@ -1,1 +1,0 @@
-export const UNKNOWN_ERROR = 'An unknown error has occurred';
